refactor: split demo app modules by target due to Kotlin 2.3.0 Android plugin change
