+++ demo/android/src/main/kotlin/com/kitakkun/jetwhale/demo/DemoApplication.kt
@@ -1,6 +1,7 @@
 package com.kitakkun.jetwhale.demo
 
 import android.app.Application
+import com.kitakkun.jetwhale.demo.shared.initializeJetWhale
 
 class DemoApplication : Application() {
     override fun onCreate() {
