+++ demo/desktop/src/main/kotlin/com/kitakkun/jetwhale/demo/main.kt
@@ -1,0 +1,13 @@
+package com.kitakkun.jetwhale.demo.com.kitakkun.jetwhale.demo
+
+import androidx.compose.ui.window.singleWindowApplication
+import com.kitakkun.jetwhale.demo.shared.App
+import com.kitakkun.jetwhale.demo.shared.initializeJetWhale
+
+fun main() {
+    initializeJetWhale()
+
+    singleWindowApplication {
+        App()
+    }
+}
