+++ demo/shared/src/commonMain/kotlin/com/kitakkun/jetwhale/demo/shared/DIModule.kt
@@ -1,4 +1,4 @@
-package com.kitakkun.jetwhale.demo
+package com.kitakkun.jetwhale.demo.shared
 
 import com.kitakkun.jetwhale.plugins.example.agent.ExampleAgentPlugin
