--- demo/src/jvmMain/kotlin/com/kitakkun/jetwhale/demo/main.kt
@@ -1,11 +1,0 @@
-package com.kitakkun.jetwhale.demo
-
-import androidx.compose.ui.window.singleWindowApplication
-
-fun main() {
-    initializeJetWhale()
-
-    singleWindowApplication {
-        App()
-    }
-}
