+++ gradle/libs.versions.toml
@@ -62,6 +62,7 @@ androidxDatastorePreferences = { module = "androidx.datastore:datastore-preferen
 
 # compose
 material3 = { module = "org.jetbrains.compose.material3:material3", version.ref = "material3" }
+jetbrainsComposeRuntime = { module = "org.jetbrains.compose.runtime:runtime", version.ref = "jetbrainsCompose" }
 rin = { module = "io.github.takahirom.rin:rin", version.ref = "rin" }
 
 # navigation3
@@ -127,6 +128,7 @@ ktorClient = [
 [plugins]
 # kotlin
 kotlinJvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
+kotlinAndroid = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 kotlinMultiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
 ksp = { id = "com.google.devtools.ksp", version.ref = "ksp" }
 kotlinDsl = { id = "org.gradle.kotlin.kotlin-dsl", version.ref = "kotlinDsl" }
